totalvalue method returns float

+++ app/src/main/java/com/example/huascar/shoppingbasket/ShoppingBasket.java
@@ -30,7 +30,7 @@ public class ShoppingBasket {
         basket.clear();
     }
 
-    public int TotalValue() {
+    public float TotalValue() {
         int total = 0;
         for (Item item : basket) {
             total += item.getPrice();
